value: fix potential precision issues with MantExp (#228)
* value: fix potential precision issues with MantExp

x.MantExp(z) sets z.prec = x.prec. As a result, using z as a return
value where it is expected to have a set precision will not always
work as intended. Another problem with this occurs in loop when
comparing zPrev.SetPrec(p).Set(z) with a future value of z and
z.prec > p. This results in an infinite loop.

* value: add test for #228

* tests: reset float precision in reset()