ivy: change indexing to have the same precedence as any other binary operator
This change addresses an ambigiuity. Without it,

	1 2 3 x[2]

changes parse depending on whether x is a vector. Now it
consistently treats

	1 2 3 x

as a vector. However, it means that

	x[1] x[2]

has new behavior, and parentheses are required to preserve
the old behavior.

The change is simple, requiring that indexing always be parsed as
part of building the operand in Parser.NumberOrVector.  But it has
minor knock-on effects in things like printing operator text. With
an excess of caution, there are more parentheses around indexing
expressions in such places now to avoid misrepresenting the
precedence.