ivy: add )clear special command
Fix a long-standing irritation. I put in "opdelete" because it was
necessary and easy but it was so ugly.  Replace it with a special
command and its own minilanguage so we can clear operators and
variables or, with no arguments, everything.

Examples:

Remove whatever is called x:
	)clear x

Remove the unary op fac:
	)clear unary fac

Remove the binary op gcd:
	)clear binary gcd

Remove all variables.
	)clear vars