ivy: remember the source text
Keep a copy of the source text as we parse. This allows us to
remember comments and formatting inside a function definition. It
also allows us to guarantee that a reparse, such as from a save
file, will give us the same result, allowing us to avoid concerns
about parentheses and precedence.

If the input base changes, a reparse of the text may make an error.
Unlikely, though, especially as we mitigate it here by remember the
ibase setting when the function is parsed, and restoring it in the
save file as needed. Not a full mitigation, but pretty good, and
of course now the user is free to see and edit the source if
necessary.

Since we have the text, allow us to see it again with the new )last
special command. This could be handy if a huge expression scrolls
the input off the screen.

Also document the long-provided but undocumented )trace special
command.