ivy: better support for unquoted file names in )get
The previous CL permitted this, but I didn't test it well enough.
A file name with a period in it caused the lexer to barf before we
got a chance to handle the input. But the failing check is redundant,
as numbers are full parsed outside the lexer, so we just let it
through and, if there is a true error, it will be caught higher up.