ivy: fix fix for failing function definition
The old code was subtly (and rarely) incorrect. Since the first
step in op definition is to install a placeholder, the order in the
definition list changes.  We need to restore the order if the
definition fails. Ugly but easy: remember where it was and put it
back there.

I took a look at slices.Insert. Don't need anything like that.