ivy: add control flow constructs for if and while
Two new control structures, introduced by keywords beginning with
a colon, inspired by Dyalog.

	:if condition; expressionList :else expressionList :end
	:while condition; expressionList :end

Both these parse as plain expressions. It is an expression
language after all. They evaluate to the last expression evaluated
in their body (or empty if nothing happens). This allows for some
nice concision, like this:

	op fac n =
		f = n
		:while 0 < n=n-1
			f = f*n
		:end

	fac 6
	720

Notice the use of assignment as an expression, and the final
multiplication yields the answer.

This took some work in the parser because for the first time we
have expressions that can span multiple lines: that :while...:end
is a single expression. It also elevates ExprList to be used more
widely, such as in the body of a while but also as the body of an
operator. This change may be a step on the way to nonce operators.

There is also a change in the meaning of the old ":" operator.
Before, it generated a value for the operator and terminated the
operator. It now does the same, but for a block; if the block is
inside an operator, the operator will continue to execute. Old code
will still work as the only block structure was an operator definition,
but now we can use it as a switch statement nestled inside another
block.

If you really want to stop the operator, use the new :ret (return)
construct. Within an operator with no if or while, it behaves just
the same as colon did and still does in that case, while allowing
us to distinguish the completion of the operator from the arrival
at a final value.

And of course although it hardly matters the colon operator is now
not restricted to operators, and parses like any expression, although
its evaluation is handled specially, as is that of :if and :while.

The parser was worked on quite a bit, and much cleanup ensued.
Probably lots of bugs added too, so I'll hold off tagging until
it's soaked a bit more. It took me many iterations to balance
everything out but I'm pleased with how the implementation finally
turned out.

I am considering adding a :begin or some such so we can have a plain
block.  For the moment, if you need one, you can always write,

	:if 1; block :end

Also found a day 1 bug in )debug parse: It crashed when printing a
unary. Yikes.

Fixes #247

Scratches long-standing itch.