ivy: add PrintValue
The addition of :while and :if led to the unwrapping of QuietValue
on return, so that a terminal assignment does not stop the value
from being printed. But this had the unintended consequence of
causing printed values to appear twice. The simple solution is to
have two kinds of quiet values, one for assignment and one for
print. Clumsy but, as I said, simple.

Fixes #250