ivy: add support for libraries
A new special command allows one to execute the contents of a
pre-loaded library file.

	)lib foo

loads (executes) library foo.  If no library name is provided, all
available libraries are listed by name.

Even if the library is not loaded,

	)lib doc foo
Shows documentation for the library foo, at the moment defined to
be the leading comment in the file.

	)lib ops foo

lists the user-defined operators in the library and

	)lib vars foo

lists its global variables.

Libraries may be added by adding an ivy file to lib/lib and running
go generate in lib.

At the moment there is only a single library, a nutty pile of code
that tracks values and units of physical constants such as mass,
velocity, and so on. That library, called units, still needs a lot
of work but that can develop separately.

Also clean up a few gofmt issues. What, gofmt does smart quotes
now? Give me a break.